+++ src/commands/config.ts
@@ -3,6 +3,7 @@ import {ChatInputCommandInteraction, EmbedBuilder, PermissionFlagsBits} from 'di
 import {injectable} from 'inversify';
 import {prisma} from '../utils/db.js';
 import Command from './index.js';
+import {getGuildSettings} from '../utils/get-guild-settings';
 
 @injectable()
 export default class implements Command {
@@ -96,11 +97,7 @@ export default class implements Command {
       case 'get': {
         const embed = new EmbedBuilder().setTitle('Config');
 
-        const config = await prisma.setting.findUnique({where: {guildId: interaction.guild!.id}});
-
-        if (!config) {
-          throw new Error('no config found');
-        }
+        const config = await getGuildSettings(interaction.guild!.id);
 
         const settingsToShow = {
           'Playlist Limit': config.playlistLimit,